Calculate residual phenotype risk scores

The residual phenotype risk score corresponds to how much a person's phenotype risk score deviates from what is expected given their characteristics (e.g. age, sex, and race)

getResidualScores(demos, scores, glmFormula, dopar = FALSE)

Arguments

demos

A data.table of covariates to be used in the calculation of residual phenotype risk score. Must have column person_id.

scores

A data.table containing the phenotype risk score for each person for each disease. Must have columns person_id, disease_id, and score.

glmFormula

A formula object representing the linear model of covariates to be used in the calculation of residual phenotype risk score. All covariates must be existing columns in demos.

dopar

Logical indicating whether to run calculations in parallel if a parallel backend is already set up, e.g., using doParallel::registerDoParallel(). Recommended to minimize runtime.

Value

A data.table containing the phenotype risk score and the residual phenotype risk score for each person for each disease.

Examples

library('data.table')

# map ICD codes to phecodes
phecodeOccurrences = getPhecodeOccurrences(icdSample)

# calculate weights
weights = getWeights(demoSample, phecodeOccurrences)

# vector of OMIM disease IDs to calculate PheRS for
diseaseIds = c(154700)

# mape diseases to phecodes
diseasePhecodeMap = mapDiseaseToPhecode()

# calculate PheRS
scores = getScores(
  demoSample, phecodeOccurrences, weights, diseasePhecodeMap[disease_id %in% diseaseIds])

# calculate residual PheRS
rscores = getResidualScores(demoSample, scores, glmFormula = as.formula(~ sex))