Map diseases entities to phecodes using HPO terms

This function takes a list of disease identifiers and returns the clinical features mapped to them as phecodes.

mapDiseaseToPhecode(
  diseaseIDs = unique(phers::diseaseHPOMap[db_name == "OMIM"]$disease_id),
  dbName = "OMIM",
  diseaseHPOMap = phers::diseaseHPOMap,
  HPOPhecodeMap = phers::HPOPhecodeMap,
  excludeDx = FALSE,
  diseaseDxPhecode = phers::diseaseDxPhecode
)

Arguments

diseaseIDs

A numeric vector of disease identifiers to be mapped

dbName

A character string of the name of database used for mapping. One of either 'OMIM', 'ORPHA', and 'DECIPHER'.

diseaseHPOMap

A data.table containing the mapping between disease entities in diseaseIDs and HPO terms. The columns are db_name, disease_id, term_id. By default uses the map included in this package.

HPOPhecodeMap

A data.table containing the mapping between HPO terms and phecodes. The columns are term_id and phecode. By default uses the map included in this package.

excludeDx

A boolean value indicating whether or not diagnostic phecodes should be removed from the disease to phecode map.

diseaseDxPhecode

A data.table containing mapping between diseases and the phecodes that represent being diagnosed with them. The columns are disease_id and dx_phecode.

Value

A data.table containing the mapping between diseases and phecodes. The columns are disease_id and phecode.