Perform a single genotype association test

Perform an association test between one variant and the phenotype risk scores corresponding to one disease.

runLinear(
  lmInput,
  glmFormula,
  modelType = c("additive", "dominant", "recessive", "genotypic"),
  diseaseId,
  snp,
  level = 0.95
)

Arguments

lmInput

A data.table of of the input to the genetic association model. Must have columns score, allele_count and any other covariates to be used in the linear regression model. allele_count should contain the number of copies of the minor allele each person has (additive model).

glmFormula

a formula object representing the linear model of covariates to be used in the genetic association test. All covariates must be existing columns in lmInput.

modelType

Coding of the genotypes. Either 'additive' or 'genotypic'

diseaseId

Numeric value of the disease ID in the association test.

snp

String value of the variant ID in the association test.

level

Confidence level used for calculating confidence intervals. Default is 0.95.

Value

A data.table containing to results of the genotype association test with the following columns. If the model fails to converge, NAs will be reported.

disease_id: disease identifier

gene: name of the gene

vid: variant identifier

beta: The beta coefficient for the association

se: The standard error for the beta coefficient

pval: The p-value for the association

ci_lower: The lower bound of the confint confidence interval

ci_upper: The upper bound of the confint confidence interval