Perform genotype association tests

Perform association tests between the phenotype risk scores and genotypes

getGeneticAssociations(
  scores,
  genotypes,
  demos,
  diseaseVariantMap,
  glmFormula,
  modelType = c("additive", "genotypic"),
  dopar = FALSE
)

Arguments

scores

A data.table of phenotype risk scores. Must have columns person_id, disease_id, score.

genotypes

A data.table of genotypes. Must have columns person_id, and variant IDs to be tested. each variant ID column should contain the number of copies of the minor allele each person has (additive model).

demos

A data.table of covariates to be used in the association analysis. Must have column person_id.

diseaseVariantMap

A data.table that maps diseases to variants. Must have columns disease_id and vid. disease_id must include all disease IDs in scores.

glmFormula

A formula object representing the linear model of covariates to be used in the genetic association tests. All covariates must be existing columns in demos.

modelType

Coding of the genotypes. Either 'additive' or 'genotypic'.

dopar

Logical indicating whether to run calculations in parallel if a parallel backend is already set up, e.g., using doParallel::registerDoParallel(). Recommended to minimize runtime.

Value

A data.table containing to results of the genotype association tests with the following columns. If the model fails to converge, NAs will be reported.

disease_id: disease identifier

gene: name of the gene

vid: variant identifier

beta: The beta coefficient for the variant

se: The standard error for the beta coefficient

p: The p-value for the variant

lower: The lower bound of the confint confidence interval

upper: The upper bound of the confint confidence interval

Examples

library('data.table')

# map ICD codes to phecodes
phecodeOccurrences = getPhecodeOccurrences(icdSample)

# calculate weights
weights = getWeights(demoSample, phecodeOccurrences)

# vector of OMIM disease IDs to calculate PheRS for
diseaseIds = c(154700)

# mape diseases to phecodes
diseasePhecodeMap = mapDiseaseToPhecode()

# calculate PheRS
scores = getScores(
  demoSample, phecodeOccurrences, weights, diseasePhecodeMap[disease_id %in% diseaseIds])

# create a map of diseases and variants
diseaseGeneMap = data.table(disease_id = 154700, gene = 'FBN1')
geneVarMap = data.table(gene = 'FBN1', vid = paste0('snp', 1:20))
diseaseVariantMap = merge(diseaseGeneMap, geneVarMap, by = 'gene')

# run genetic association tests
genoStats = getGeneticAssociations(
  scores, genoSample, demoSample, diseaseVariantMap, glmFormula = as.formula(~ sex),
  modelType = 'additive')